refactor(chat): AskUserQuestion type safety and model organization
- Move UserQuestion and QuestionOption to chat_message.dart where other
  chat data models live; user_question_card.dart becomes a re-export shim
  so existing imports continue to work without changes
- Pre-build tool_result index in toMessages() for O(1) lookups instead
  of the previous O(N*K) linear scan per tool_use block; removes the now-
  unused _findToolResult() helper

Closes #87
