+++ app/lib/features/chat/models/chat_message.dart
@@ -1,3 +1,50 @@
+/// A parsed question from an AskUserQuestion tool call.
+///
+/// Typed view over the raw [Map<String, dynamic>] stored in [UserQuestionData].
+class UserQuestion {
+  final String question;
+  final String header;
+  final List<QuestionOption> options;
+  final bool multiSelect;
+
+  const UserQuestion({
+    required this.question,
+    required this.header,
+    required this.options,
+    this.multiSelect = false,
+  });
+
+  factory UserQuestion.fromJson(Map<String, dynamic> json) {
+    final optionsList = (json['options'] as List<dynamic>?) ?? [];
+    return UserQuestion(
+      question: json['question'] as String? ?? '',
+      header: json['header'] as String? ?? '',
+      options: optionsList
+          .map((o) => QuestionOption.fromJson(o as Map<String, dynamic>))
+          .toList(),
+      multiSelect: json['multiSelect'] as bool? ?? false,
+    );
+  }
+}
+
+/// A single option for an [UserQuestion].
+class QuestionOption {
+  final String label;
+  final String description;
+
+  const QuestionOption({
+    required this.label,
+    required this.description,
+  });
+
+  factory QuestionOption.fromJson(Map<String, dynamic> json) {
+    return QuestionOption(
+      label: json['label'] as String? ?? '',
+      description: json['description'] as String? ?? '',
+    );
+  }
+}
+
 /// Role of the message sender
 enum MessageRole { user, assistant }
